refactor(mcp): address kill-switch PR review feedback
